refactored latin stemmer package
